Customer & Order Analytics Project
How many orders were placed in January?
How many of those orders were for an iPhone?
Select the customer account numbers for all the orders that were placed in February.
Which product was the cheapest one sold in January, and what was the price?
What is the total revenue for each product sold in January? (Revenue can be calculated using the number of products sold and the price of the products).
Which products were sold in February at 548 Lincoln St, Seattle, WA 98101, how many of each were sold, and what was the total revenue?
How many customers ordered more than 2 products at a time in February, and what was the average amount spent for those customers?